Fix ambiguity of `clamp` use, prefer int version

+++ src/burst.cpp
@@ -103,7 +103,7 @@ void Burst::process(const ProcessArgs &args) {
 	float randomDelta = 0;
 
 	timeParam = clamp(params[TIME_PARAM].getValue() + (params[TIME_ATT_PARAM].getValue() * inputs[TIME_INPUT].getVoltage() / 10.0 * MAX_TIME), 0.0f, MAX_TIME);
-	pulseParam = clamp((int)params[REP_PARAM].getValue() + (inputs[REP_INPUT].getVoltage() * params[REP_ATT_PARAM].getValue() / 10.0 * MAX_REPS), 0, MAX_REPS);
+	pulseParam = clamp((int)params[REP_PARAM].getValue() + (int)(inputs[REP_INPUT].getVoltage() * params[REP_ATT_PARAM].getValue() / 10.0 * MAX_REPS), 0, MAX_REPS);
 
 	//exponential scaling for timeparam
 	timeParam = (exp(timeParam) - 1) / (euler - 1);
